pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "App"
readme = "README.md"
authors = [{ name = "Ronaldo Olivares", email = "ronaldoolivarespalacios@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.6",
    "logfire[fastapi,fastapihttpx,httpx,openai]>=4.3.5",
    "pydantic-settings>=2.10.1",
    "scalar-fastapi>=1.2.3",
    "supabase>=2.18.1",
    "uvicorn>=0.37.0",
]
[dependency-groups]
dev = [
    "langgraph-api>=0.3.0",
    "langgraph-cli>=0.3.7",
    "mypy>=1.17.1",
    "ruff>=0.12.9",
]

[tool.mypy]
strict = true
warn_unused_ignores = false
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py313"
exclude = []

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"


[tool.uv.build-backend]
module-root = ""
